TA-224: Refactor ExUi approve org provider tests
