+++ src/contractTest/java/uk/gov/hmcts/reform/professionalapi/provider/WebMvcProviderTest.java
@@ -1,0 +1,23 @@
+package uk.gov.hmcts.reform.professionalapi.provider;
+
+import au.com.dius.pact.provider.junit5.PactVerificationContext;
+import au.com.dius.pact.provider.spring.junit5.MockMvcTestTarget;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.test.web.servlet.MockMvc;
+
+@AutoConfigureMockMvc(addFilters = false)
+public class WebMvcProviderTest extends BaseProviderTest {
+
+    @Autowired
+    MockMvc mockMvc;
+
+    @BeforeEach
+    void before(PactVerificationContext context) {
+        //System.getProperties().setProperty("pact.verifier.publishResults", "true");
+        if (context != null) {
+            context.setTarget(new MockMvcTestTarget(mockMvc));
+        }
+    }
+}
