+++ src/contractTest/resources/application-contract.yaml
@@ -35,10 +35,6 @@ postgres.password: ${FUNC_DATABASE_PASS}
 postgres.port: ${FUNC_DATABASE_PORT}
 #######
 
-logging:
-  level:
-    org.springframework.web: DEBUG
-    au.com.dius.pact: DEBUG
